feat: implement react-router and changes in sctructure
