+++ src/app/app.scss
@@ -7,7 +7,3 @@
 html,body{
     font-size: 20px;
 }
-
-.container{
-    display: flex;
-}
