+++ src/containers/dashboard/index.js
@@ -1,0 +1,10 @@
+import React from "react";
+import "./dashboard.scss";
+
+class Dashboard extends React.Component {
+  render() {
+    return <div></div>;
+  }
+}
+
+export default Dashboard;
