+++ src/containers/login/login.scss
@@ -1,0 +1,3 @@
+.container{
+    display: flex;
+}
