+++ src/styles/colors.scss
@@ -11,6 +11,7 @@ $input-placeholder-moodle: #c4c4c4;
 // -----COLORS-----
 $btn-bg-moodle: #0C7277;
 $btn-bg-trello: #68E45D;
+$dashboard-primary: #FDF5BF;
 
 // -----COLORS-----
